ci: update Storybook repros repo only on changes
Currently, every night repros for most of the frameworks are recreated and pushed to: https://github.com/storybookjs/repro-templates
This is done for both `main` and `next` branches, and the full repo is recreated and force pushed after each run.

This commit improves a bit the behavior by committing and pushing only the differences between the current run and the previous one.

This offers a nice Git history allowing you to easily check out old versions of the repros.